Fix an off_t, Add key event framework

+++ src/hi_diff.h
@@ -64,6 +64,6 @@ typedef struct hi_diff_hunk
 hi_diff *hi_diff_calculate(hi_file *src, hi_file *dst);
 hi_diff_hunk *hi_diff_get_hunk(hi_diff *diff,
                                hi_file *file,
-                               int pos);
+                               off_t pos);
 
 #endif
